Adding a lot

+++ src/main/java/com/github/mrgeotech/Shape.java
@@ -2,7 +2,7 @@ package com.github.mrgeotech;
 
 import org.bukkit.Bukkit;
 
-public record Shape(ParticleHandler.Type type, Runnable runnable, int x, int y, IntegerContainer taskID) {
+public record Shape(ParticleHandler.Type type, Runnable runnable, double x, double y, IntegerContainer taskID) {
 
     public void run() {
         taskID.integer = Bukkit.getScheduler()
